tidy up, split scripts, adds README
