+++ src/cpg_seqr_loader/first_workflow.py
@@ -1,0 +1,23 @@
+#!/usr/bin/env python3
+
+"""
+This is the entry point for the first workflow, running from gVCFs, through VDS, to a dense MatrixTable and VCFs
+"""
+
+import argparse
+
+from cpg_flow import workflow
+
+from cpg_seqr_loader.stages import CreateDenseMtFromVdsWithHail
+
+
+def cli_main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--dry_run', action='store_true', help='Dry run')
+    args = parser.parse_args()
+
+    workflow.run_workflow(stages=[CreateDenseMtFromVdsWithHail], dry_run=args.dry_run)
+
+
+if __name__ == '__main__':
+    cli_main()
